backend/cloud-function/requirements.txt
functions-framework
pymysql
google-auth
DateTime
mysql.connector
SQLAlchemy==1.3.12
google-cloud-secret-manager